Extracted `sql` subproject.
This also moves more of SQL² into the `quasar.sql` package and renames
the confusing `main` subproject to `interface`.
