+++ CHANGELOG/revision.md
@@ -1,0 +1,1 @@
+- extract a `sql` subproject
