+++ interface/src/main/scala/quasar/main/package.scala
