+++ interface/src/test/scala/quasar/config/CoreConfigSpec.scala
